refactor: aplicando estrutura de projeto DDD
