+++ contextos_de_negocios/dominio/entidades/seguranca.py
