+++ contextos_de_negocios/dominio/entidades/usuario.py
