+++ contextos_de_negocios/dominio/objetos_de_valor/transacao_bancaria.py
