+++ contextos_de_negocios/main.py
@@ -6,13 +6,21 @@ import infra.database
 from contextos_de_negocios.utils.constantes import SENTRY_DSN
 from infra.database import engine, criar_primeiro_usuario
 
-from contextos_de_negocios.servicos.routes import router as servicos_router
-from contextos_de_negocios.conta_bancaria.routes import router as conta_bancaria_router
-from contextos_de_negocios.transacao_bancaria.routes import (
+from contextos_de_negocios.pontos_de_entrada.api_publica.recursos_seguranca import (
+    router as servicos_router,
+)
+from contextos_de_negocios.pontos_de_entrada.api_publica.recursos_conta_bancaria import (
+    router as conta_bancaria_router,
+)
+from contextos_de_negocios.pontos_de_entrada.api_publica.recursos_transacao_bancaria import (
     router as transacao_bancaria_router,
 )
-from contextos_de_negocios.cliente.routes import router as cliente_router
-from contextos_de_negocios.usuario.routes import router as usuario_router
+from contextos_de_negocios.pontos_de_entrada.api_publica.recursos_cliente import (
+    router as cliente_router,
+)
+from contextos_de_negocios.pontos_de_entrada.api_publica.recursos_usuario import (
+    router as usuario_router,
+)
 
 sentry_sdk.init(
     dsn=SENTRY_DSN,
