+++ infra/database.py
@@ -49,9 +49,13 @@ async def extrair_sessao(conexao: AsyncIterable = get_db()) -> AsyncSession:
 
 
 async def criar_primeiro_usuario() -> None:
-    from contextos_de_negocios.usuario.repositorio import RepoUsuarioLeitura
-    from contextos_de_negocios.usuario.schemas import CadastrarEAtualizarUsuario
-    from contextos_de_negocios.usuario.controllers import UsuarioControllers
+    from contextos_de_negocios.repositorio.repo_consulta.usuario import (
+        RepoUsuarioLeitura,
+    )
+    from contextos_de_negocios.dominio.entidades.usuario import (
+        CadastrarEAtualizarUsuario,
+    )
+    from contextos_de_negocios.servicos.executores.usuario import UsuarioControllers
 
     async with SessionLocal() as session:
         usuarios = await RepoUsuarioLeitura.consultar_todos(session=session)
